Restructure code to follow one-pub-per-file convention
Refactored the codebase to follow the "one pub per file" convention as
specified in CLAUDE.md. Each public item now has its own file, with
module directories serving as table-of-contents.

Changes:
- Restructured select.rs into select/ module (9 public items)
- Restructured iter.rs into iter/ module (9 public items)
- Restructured tree.rs into tree/ module (6 public items)
- Restructured attributes.rs into attributes/ module (3 public items)
- Added justification comment to cell_extras.rs for grouped traits
- Added module_inception = "forbid" lint to Cargo.toml
- Renamed attribute.rs → attrib.rs and attributes.rs → attribs.rs
  to avoid module_inception warning

Benefits:
- Clear file-to-type mapping for easy navigation
- Better git history (changes to specific types touch only their files)
- Reduced merge conflicts
- Test colocation maintained
- 100% API backward compatibility via re-exports

All tests pass (44 unit + 22 doc tests)
No clippy warnings
Code coverage: 72.19%

Resolves #30
