+++ src/iter/node_edge.rs
@@ -1,0 +1,13 @@
+/// Marks either the start or the end of a node.
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum NodeEdge<T> {
+    /// Indicates that start of a node that has children.
+    /// Yielded by `Traverse::next` before the node's descendants.
+    /// In HTML or XML, this corresponds to an opening tag like `<div>`
+    Start(T),
+
+    /// Indicates that end of a node that has children.
+    /// Yielded by `Traverse::next` after the node's descendants.
+    /// In HTML or XML, this corresponds to a closing tag like `</div>`
+    End(T),
+}
