--- src/select.rs
@@ -1,876 +1,0 @@
-// Addressing this lint is a semver-breaking change.
-// Remove this once the issue has been addressed.
-#![allow(clippy::result_unit_err)]
-
-use crate::attributes::ExpandedName;
-use crate::iter::{NodeIterator, Select};
-use crate::node_data_ref::NodeDataRef;
-use crate::tree::{ElementData, Node, NodeData, NodeRef};
-use cssparser::{self, CowRcStr, ParseError, SourceLocation, ToCss};
-use html5ever::{LocalName, Namespace};
-use precomputed_hash::PrecomputedHash;
-use selectors::attr::{AttrSelectorOperation, CaseSensitivity, NamespaceConstraint};
-use selectors::context::QuirksMode;
-use selectors::parser::SelectorParseErrorKind;
-use selectors::parser::{
-    NonTSPseudoClass, Parser, Selector as GenericSelector, SelectorImpl, SelectorList,
-};
-use selectors::{self, matching, OpaqueElement};
-use std::borrow::Borrow;
-use std::fmt;
-use std::ops::Deref;
-
-/// The definition of whitespace per CSS Selectors Level 3 § 4.
-///
-/// Copied from rust-selectors.
-static SELECTOR_WHITESPACE: &[char] = &[' ', '\t', '\n', '\r', '\x0C'];
-
-/// Context for compiling CSS selectors.
-///
-/// This struct holds configuration that affects how selectors are parsed and matched,
-/// including namespace prefix mappings for namespace-aware selector matching.
-///
-/// # Examples
-///
-/// ```
-/// use brik::SelectorContext;
-/// use html5ever::ns;
-///
-/// let mut context = SelectorContext::new();
-/// context.add_namespace("svg".to_string(), ns!(svg));
-/// context.set_default_namespace(ns!(html));
-/// ```
-#[derive(Clone, Debug, Default)]
-pub struct SelectorContext {
-    /// Map from namespace prefixes to namespace URIs.
-    namespaces: std::collections::HashMap<String, Namespace>,
-    /// Optional default namespace for unprefixed element selectors.
-    default_namespace: Option<Namespace>,
-}
-
-impl SelectorContext {
-    /// Create a new empty selector context.
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Add a namespace prefix mapping.
-    ///
-    /// This allows selectors to use the prefix in type selectors (e.g., `svg|rect`)
-    /// and attribute selectors (e.g., `[tmpl|if]`).
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use brik::SelectorContext;
-    /// use html5ever::ns;
-    ///
-    /// let mut context = SelectorContext::new();
-    /// context.add_namespace("svg".to_string(), ns!(svg));
-    /// ```
-    pub fn add_namespace(&mut self, prefix: String, url: Namespace) -> &mut Self {
-        self.namespaces.insert(prefix, url);
-        self
-    }
-
-    /// Set the default namespace for unprefixed element selectors.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use brik::SelectorContext;
-    /// use html5ever::ns;
-    ///
-    /// let mut context = SelectorContext::new();
-    /// context.set_default_namespace(ns!(html));
-    /// ```
-    pub fn set_default_namespace(&mut self, url: Namespace) -> &mut Self {
-        self.default_namespace = Some(url);
-        self
-    }
-}
-
-/// Wrapper for String that implements ToCss
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub struct AttrValue(String);
-
-impl ToCss for AttrValue {
-    fn to_css<W>(&self, dest: &mut W) -> fmt::Result
-    where
-        W: fmt::Write,
-    {
-        cssparser::serialize_string(&self.0, dest)
-    }
-}
-
-impl From<String> for AttrValue {
-    fn from(s: String) -> Self {
-        AttrValue(s)
-    }
-}
-
-impl<'a> From<&'a str> for AttrValue {
-    fn from(s: &'a str) -> Self {
-        AttrValue(s.to_string())
-    }
-}
-
-impl Deref for AttrValue {
-    type Target = String;
-    fn deref(&self) -> &String {
-        &self.0
-    }
-}
-
-impl Borrow<str> for AttrValue {
-    fn borrow(&self) -> &str {
-        &self.0
-    }
-}
-
-impl AsRef<str> for AttrValue {
-    fn as_ref(&self) -> &str {
-        &self.0
-    }
-}
-
-/// Wrapper for LocalName that implements ToCss
-#[derive(Debug, Clone, Eq, PartialEq, Hash, Default)]
-pub struct LocalNameSelector(LocalName);
-
-impl ToCss for LocalNameSelector {
-    fn to_css<W>(&self, dest: &mut W) -> fmt::Result
-    where
-        W: fmt::Write,
-    {
-        cssparser::serialize_identifier(&self.0, dest)
-    }
-}
-
-impl From<LocalName> for LocalNameSelector {
-    fn from(name: LocalName) -> Self {
-        LocalNameSelector(name)
-    }
-}
-
-impl<'a> From<&'a str> for LocalNameSelector {
-    fn from(s: &'a str) -> Self {
-        LocalNameSelector(LocalName::from(s))
-    }
-}
-
-impl Deref for LocalNameSelector {
-    type Target = LocalName;
-    fn deref(&self) -> &LocalName {
-        &self.0
-    }
-}
-
-impl Borrow<LocalName> for LocalNameSelector {
-    fn borrow(&self) -> &LocalName {
-        &self.0
-    }
-}
-
-impl AsRef<LocalName> for LocalNameSelector {
-    fn as_ref(&self) -> &LocalName {
-        &self.0
-    }
-}
-
-impl PrecomputedHash for LocalNameSelector {
-    fn precomputed_hash(&self) -> u32 {
-        self.0.precomputed_hash()
-    }
-}
-
-/// Selector implementation for Brik's DOM.
-#[derive(Debug, Clone)]
-pub struct BrikSelectors;
-
-impl SelectorImpl for BrikSelectors {
-    type AttrValue = AttrValue;
-    type Identifier = LocalNameSelector;
-    type LocalName = LocalNameSelector;
-    type NamespacePrefix = LocalNameSelector;
-    type NamespaceUrl = Namespace;
-    type BorrowedNamespaceUrl = Namespace;
-    type BorrowedLocalName = LocalName;
-
-    type NonTSPseudoClass = PseudoClass;
-    type PseudoElement = PseudoElement;
-
-    type ExtraMatchingData<'a> = ();
-}
-
-/// Parser for CSS selectors.
-struct BrikParser<'a> {
-    /// Selector context containing namespace mappings and other configuration.
-    context: &'a SelectorContext,
-}
-
-impl<'a> BrikParser<'a> {
-    /// Create a new parser with the given selector context.
-    fn new(context: &'a SelectorContext) -> Self {
-        BrikParser { context }
-    }
-}
-
-impl<'i, 'a> Parser<'i> for BrikParser<'a> {
-    type Impl = BrikSelectors;
-    type Error = SelectorParseErrorKind<'i>;
-
-    fn parse_non_ts_pseudo_class(
-        &self,
-        location: SourceLocation,
-        name: CowRcStr<'i>,
-    ) -> Result<PseudoClass, ParseError<'i, SelectorParseErrorKind<'i>>> {
-        use self::PseudoClass::*;
-        if name.eq_ignore_ascii_case("any-link") {
-            Ok(AnyLink)
-        } else if name.eq_ignore_ascii_case("link") {
-            Ok(Link)
-        } else if name.eq_ignore_ascii_case("visited") {
-            Ok(Visited)
-        } else if name.eq_ignore_ascii_case("active") {
-            Ok(Active)
-        } else if name.eq_ignore_ascii_case("focus") {
-            Ok(Focus)
-        } else if name.eq_ignore_ascii_case("hover") {
-            Ok(Hover)
-        } else if name.eq_ignore_ascii_case("enabled") {
-            Ok(Enabled)
-        } else if name.eq_ignore_ascii_case("disabled") {
-            Ok(Disabled)
-        } else if name.eq_ignore_ascii_case("checked") {
-            Ok(Checked)
-        } else if name.eq_ignore_ascii_case("indeterminate") {
-            Ok(Indeterminate)
-        } else {
-            Err(
-                location.new_custom_error(SelectorParseErrorKind::UnsupportedPseudoClassOrElement(
-                    name,
-                )),
-            )
-        }
-    }
-
-    fn default_namespace(&self) -> Option<Namespace> {
-        self.context.default_namespace.clone()
-    }
-
-    fn namespace_for_prefix(&self, prefix: &LocalNameSelector) -> Option<Namespace> {
-        self.context
-            .namespaces
-            .get(prefix.as_ref().as_ref())
-            .cloned()
-    }
-}
-
-/// Supported CSS pseudo-classes for element matching.
-#[derive(PartialEq, Eq, Clone, Debug, Hash)]
-pub enum PseudoClass {
-    /// Matches `:any-link` (any link element).
-    AnyLink,
-    /// Matches `:link` (unvisited links).
-    Link,
-    /// Matches `:visited` (visited links).
-    Visited,
-    /// Matches `:active` (activated elements).
-    Active,
-    /// Matches `:focus` (focused elements).
-    Focus,
-    /// Matches `:hover` (hovered elements).
-    Hover,
-    /// Matches `:enabled` (enabled form elements).
-    Enabled,
-    /// Matches `:disabled` (disabled form elements).
-    Disabled,
-    /// Matches `:checked` (checked form elements).
-    Checked,
-    /// Matches `:indeterminate` (indeterminate form elements).
-    Indeterminate,
-}
-
-impl NonTSPseudoClass for PseudoClass {
-    type Impl = BrikSelectors;
-
-    fn is_active_or_hover(&self) -> bool {
-        matches!(*self, PseudoClass::Active | PseudoClass::Hover)
-    }
-
-    fn is_user_action_state(&self) -> bool {
-        matches!(
-            *self,
-            PseudoClass::Active | PseudoClass::Hover | PseudoClass::Focus
-        )
-    }
-}
-
-impl ToCss for PseudoClass {
-    fn to_css<W>(&self, dest: &mut W) -> fmt::Result
-    where
-        W: fmt::Write,
-    {
-        dest.write_str(match *self {
-            PseudoClass::AnyLink => ":any-link",
-            PseudoClass::Link => ":link",
-            PseudoClass::Visited => ":visited",
-            PseudoClass::Active => ":active",
-            PseudoClass::Focus => ":focus",
-            PseudoClass::Hover => ":hover",
-            PseudoClass::Enabled => ":enabled",
-            PseudoClass::Disabled => ":disabled",
-            PseudoClass::Checked => ":checked",
-            PseudoClass::Indeterminate => ":indeterminate",
-        })
-    }
-}
-
-/// CSS pseudo-elements (currently none are supported).
-#[derive(PartialEq, Eq, Clone, Debug, Hash)]
-pub enum PseudoElement {}
-
-impl ToCss for PseudoElement {
-    fn to_css<W>(&self, _dest: &mut W) -> fmt::Result
-    where
-        W: fmt::Write,
-    {
-        match *self {}
-    }
-}
-
-impl selectors::parser::PseudoElement for PseudoElement {
-    type Impl = BrikSelectors;
-}
-
-impl selectors::Element for NodeDataRef<ElementData> {
-    type Impl = BrikSelectors;
-
-    #[inline]
-    fn opaque(&self) -> OpaqueElement {
-        let node: &Node = self.as_node();
-        OpaqueElement::new(node)
-    }
-
-    #[inline]
-    fn is_html_slot_element(&self) -> bool {
-        false
-    }
-    #[inline]
-    fn parent_node_is_shadow_root(&self) -> bool {
-        false
-    }
-    #[inline]
-    fn containing_shadow_host(&self) -> Option<Self> {
-        None
-    }
-
-    #[inline]
-    fn parent_element(&self) -> Option<Self> {
-        self.as_node().parent().and_then(NodeRef::into_element_ref)
-    }
-    #[inline]
-    fn prev_sibling_element(&self) -> Option<Self> {
-        self.as_node().preceding_siblings().elements().next()
-    }
-    #[inline]
-    fn next_sibling_element(&self) -> Option<Self> {
-        self.as_node().following_siblings().elements().next()
-    }
-    #[inline]
-    fn first_element_child(&self) -> Option<Self> {
-        self.as_node().children().elements().next()
-    }
-    #[inline]
-    fn is_empty(&self) -> bool {
-        self.as_node().children().all(|child| match *child.data() {
-            NodeData::Element(_) => false,
-            NodeData::Text(ref text) => text.borrow().is_empty(),
-            _ => true,
-        })
-    }
-    #[inline]
-    fn is_root(&self) -> bool {
-        match self.as_node().parent() {
-            None => false,
-            Some(parent) => matches!(*parent.data(), NodeData::Document(_)),
-        }
-    }
-
-    #[inline]
-    fn is_html_element_in_html_document(&self) -> bool {
-        // FIXME: Have a notion of HTML document v.s. XML document?
-        self.name.ns == ns!(html)
-    }
-
-    #[inline]
-    fn has_local_name(&self, name: &LocalName) -> bool {
-        self.name.local == *name
-    }
-    #[inline]
-    fn has_namespace(&self, namespace: &Namespace) -> bool {
-        self.name.ns == *namespace
-    }
-
-    #[inline]
-    fn is_part(&self, _name: &LocalNameSelector) -> bool {
-        false
-    }
-
-    #[inline]
-    fn imported_part(&self, _: &LocalNameSelector) -> Option<LocalNameSelector> {
-        None
-    }
-
-    #[inline]
-    fn is_pseudo_element(&self) -> bool {
-        false
-    }
-
-    #[inline]
-    fn is_same_type(&self, other: &Self) -> bool {
-        self.name == other.name
-    }
-
-    #[inline]
-    fn is_link(&self) -> bool {
-        self.name.ns == ns!(html)
-            && matches!(
-                self.name.local,
-                local_name!("a") | local_name!("area") | local_name!("link")
-            )
-            && self
-                .attributes
-                .borrow()
-                .map
-                .contains_key(&ExpandedName::new(ns!(), local_name!("href")))
-    }
-
-    #[inline]
-    fn has_id(&self, id: &LocalNameSelector, case_sensitivity: CaseSensitivity) -> bool {
-        self.attributes
-            .borrow()
-            .get(local_name!("id"))
-            .is_some_and(|id_attr| case_sensitivity.eq(id.as_bytes(), id_attr.as_bytes()))
-    }
-
-    #[inline]
-    fn has_class(&self, name: &LocalNameSelector, case_sensitivity: CaseSensitivity) -> bool {
-        let name = name.as_bytes();
-        !name.is_empty()
-            && if let Some(class_attr) = self.attributes.borrow().get(local_name!("class")) {
-                class_attr
-                    .split(SELECTOR_WHITESPACE)
-                    .any(|class| case_sensitivity.eq(class.as_bytes(), name))
-            } else {
-                false
-            }
-    }
-
-    #[inline]
-    fn attr_matches(
-        &self,
-        ns: &NamespaceConstraint<&Namespace>,
-        local_name: &LocalNameSelector,
-        operation: &AttrSelectorOperation<&AttrValue>,
-    ) -> bool {
-        let attrs = self.attributes.borrow();
-        match *ns {
-            NamespaceConstraint::Any => attrs.map.iter().any(|(name, attr)| {
-                name.local == **local_name && operation.eval_str(attr.value.as_str())
-            }),
-            NamespaceConstraint::Specific(ns_url) => attrs
-                .map
-                .get(&ExpandedName::new(ns_url, (**local_name).clone()))
-                .is_some_and(|attr| operation.eval_str(attr.value.as_str())),
-        }
-    }
-
-    fn match_pseudo_element(
-        &self,
-        pseudo: &PseudoElement,
-        _context: &mut matching::MatchingContext<BrikSelectors>,
-    ) -> bool {
-        match *pseudo {}
-    }
-
-    fn match_non_ts_pseudo_class(
-        &self,
-        pseudo: &PseudoClass,
-        _context: &mut matching::MatchingContext<BrikSelectors>,
-    ) -> bool {
-        use self::PseudoClass::*;
-        match *pseudo {
-            Active | Focus | Hover | Enabled | Disabled | Checked | Indeterminate | Visited => {
-                false
-            }
-            AnyLink | Link => {
-                self.name.ns == ns!(html)
-                    && matches!(
-                        self.name.local,
-                        local_name!("a") | local_name!("area") | local_name!("link")
-                    )
-                    && self.attributes.borrow().contains(local_name!("href"))
-            }
-        }
-    }
-
-    #[inline]
-    fn apply_selector_flags(&self, _flags: matching::ElementSelectorFlags) {
-        // No-op for static DOM
-    }
-
-    #[inline]
-    fn has_custom_state(&self, _name: &LocalNameSelector) -> bool {
-        // Brik is a static DOM, no custom states
-        false
-    }
-
-    #[inline]
-    fn add_element_unique_hashes(&self, filter: &mut selectors::bloom::BloomFilter) -> bool {
-        let _ = filter; // Silence unused warning
-        false
-    }
-}
-
-/// A pre-compiled list of CSS Selectors.
-pub struct Selectors(pub Vec<Selector>);
-
-/// A pre-compiled CSS Selector.
-pub struct Selector(GenericSelector<BrikSelectors>);
-
-/// The specificity of a selector.
-///
-/// Opaque, but ordered.
-///
-/// Determines precedence in the cascading algorithm.
-/// When equal, a rule later in source order takes precedence.
-#[derive(Copy, Clone, Hash, Eq, PartialEq, Ord, PartialOrd)]
-pub struct Specificity(u32);
-
-impl Selectors {
-    /// Compile a list of selectors. This may fail on syntax errors or unsupported selectors.
-    ///
-    /// # Errors
-    ///
-    /// Returns `Err(())` if the selector string contains syntax errors or unsupported selectors.
-    #[inline]
-    pub fn compile(s: &str) -> Result<Selectors, ()> {
-        let context = SelectorContext::default();
-        Self::compile_with_context(s, &context)
-    }
-
-    /// Compile a list of selectors with a selector context.
-    ///
-    /// This method allows selectors to use namespace prefixes in both type selectors
-    /// (e.g., `svg|rect`) and attribute selectors (e.g., `[tmpl|if]`).
-    ///
-    /// This is the recommended method when using namespace-aware selectors.
-    ///
-    /// # Arguments
-    ///
-    /// * `s` - The selector string to compile
-    /// * `context` - A selector context containing namespace mappings
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use brik::{Selectors, SelectorContext};
-    /// use html5ever::ns;
-    ///
-    /// let mut context = SelectorContext::new();
-    /// context.add_namespace("svg".to_string(), ns!(svg));
-    ///
-    /// // Select SVG rect elements
-    /// let selectors = Selectors::compile_with_context("svg|rect", &context).unwrap();
-    /// ```
-    ///
-    /// # Errors
-    ///
-    /// Returns `Err(())` if the selector string contains syntax errors, unsupported selectors,
-    /// or references undefined namespace prefixes.
-    #[inline]
-    pub fn compile_with_context(s: &str, context: &SelectorContext) -> Result<Selectors, ()> {
-        let mut input = cssparser::ParserInput::new(s);
-        match SelectorList::parse(
-            &BrikParser::new(context),
-            &mut cssparser::Parser::new(&mut input),
-            selectors::parser::ParseRelative::No,
-        ) {
-            Ok(list) => Ok(Selectors(
-                list.slice().iter().cloned().map(Selector).collect(),
-            )),
-            Err(_) => Err(()),
-        }
-    }
-
-    /// Returns whether the given element matches this list of selectors.
-    #[inline]
-    pub fn matches(&self, element: &NodeDataRef<ElementData>) -> bool {
-        self.0.iter().any(|s| s.matches(element))
-    }
-
-    /// Filter an element iterator, yielding those matching this list of selectors.
-    #[inline]
-    pub fn filter<I>(&self, iter: I) -> Select<I, &Selectors>
-    where
-        I: Iterator<Item = NodeDataRef<ElementData>>,
-    {
-        Select {
-            iter,
-            selectors: self,
-        }
-    }
-}
-
-impl Selector {
-    /// Returns whether the given element matches this selector.
-    #[inline]
-    pub fn matches(&self, element: &NodeDataRef<ElementData>) -> bool {
-        let mut selector_caches = matching::SelectorCaches::default();
-        let mut context = matching::MatchingContext::new(
-            matching::MatchingMode::Normal,
-            None,
-            &mut selector_caches,
-            QuirksMode::NoQuirks,
-            matching::NeedsSelectorFlags::No,
-            matching::MatchingForInvalidation::No,
-        );
-        matching::matches_selector(&self.0, 0, None, element, &mut context)
-    }
-
-    /// Return the specificity of this selector.
-    pub fn specificity(&self) -> Specificity {
-        Specificity(self.0.specificity())
-    }
-}
-
-impl ::std::str::FromStr for Selectors {
-    type Err = ();
-    #[inline]
-    fn from_str(s: &str) -> Result<Selectors, ()> {
-        Selectors::compile(s)
-    }
-}
-
-impl fmt::Display for Selector {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        self.0.to_css(f)
-    }
-}
-
-impl fmt::Display for Selectors {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let mut iter = self.0.iter();
-        let first = iter
-            .next()
-            .expect("Empty Selectors, should contain at least one selector");
-        first.0.to_css(f)?;
-        for selector in iter {
-            f.write_str(", ")?;
-            selector.0.to_css(f)?;
-        }
-        Ok(())
-    }
-}
-
-impl fmt::Debug for Selector {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        fmt::Display::fmt(self, f)
-    }
-}
-
-impl fmt::Debug for Selectors {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        fmt::Display::fmt(self, f)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::parser::parse_html;
-    use crate::traits::*;
-
-    #[test]
-    fn namespace_type_selector() {
-        let html = r#"<!DOCTYPE html>
-<html>
-<body>
-    <div>HTML div</div>
-    <svg xmlns="http://www.w3.org/2000/svg">
-        <rect width="100" height="100"/>
-        <circle cx="50" cy="50" r="40"/>
-    </svg>
-</body>
-</html>"#;
-
-        let document = parse_html().one(html);
-
-        // Create context with SVG namespace
-        let mut context = SelectorContext::new();
-        context.add_namespace("svg".to_string(), ns!(svg));
-
-        // Select SVG rect elements using namespace selector
-        let selectors = Selectors::compile_with_context("svg|rect", &context).unwrap();
-        let rects = selectors
-            .filter(document.descendants().elements())
-            .collect::<Vec<_>>();
-        assert_eq!(rects.len(), 1);
-        assert_eq!(rects[0].name.local, local_name!("rect"));
-        assert_eq!(rects[0].name.ns, ns!(svg));
-
-        // Select all SVG elements using namespace wildcard
-        let selectors = Selectors::compile_with_context("svg|*", &context).unwrap();
-        let svg_elements = selectors
-            .filter(document.descendants().elements())
-            .collect::<Vec<_>>();
-        assert_eq!(svg_elements.len(), 3); // svg, rect, circle
-    }
-
-    #[test]
-    fn namespace_attribute_selector() {
-        let html = r##"<!DOCTYPE html>
-<html xmlns:custom="http://example.com/custom">
-<body>
-    <div>Regular div without custom attribute</div>
-    <div custom:attr="value">Div with custom:attr</div>
-    <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
-        <use xlink:href="#icon"/>
-    </svg>
-</body>
-</html>"##;
-
-        let document = parse_html().one(html);
-
-        // Create context with xlink namespace
-        let mut context = SelectorContext::new();
-        context.add_namespace(
-            "xlink".to_string(),
-            Namespace::from("http://www.w3.org/1999/xlink"),
-        );
-
-        // Select elements with xlink:href attribute
-        let selectors = Selectors::compile_with_context("[xlink|href]", &context).unwrap();
-        let elements = selectors
-            .filter(document.descendants().elements())
-            .collect::<Vec<_>>();
-        assert_eq!(elements.len(), 1);
-        assert_eq!(elements[0].name.local, local_name!("use"));
-    }
-
-    #[test]
-    fn namespace_selector_undefined_prefix() {
-        let context = SelectorContext::new(); // Empty context
-
-        // Try to compile selector with undefined namespace prefix
-        let result = Selectors::compile_with_context("undefined|div", &context);
-        assert!(
-            result.is_err(),
-            "Should fail with undefined namespace prefix"
-        );
-    }
-
-    #[test]
-    fn namespace_selector_backward_compatibility() {
-        let html = r#"<div class="test">Content</div>"#;
-        let document = parse_html().one(html);
-
-        // Old compile() method should still work
-        let selectors = Selectors::compile("div.test").unwrap();
-        let elements = selectors
-            .filter(document.descendants().elements())
-            .collect::<Vec<_>>();
-        assert_eq!(elements.len(), 1);
-    }
-
-    #[test]
-    fn namespace_context_builder_pattern() {
-        let mut context = SelectorContext::new();
-        context
-            .add_namespace("svg".to_string(), ns!(svg))
-            .add_namespace("html".to_string(), ns!(html))
-            .set_default_namespace(ns!(html));
-
-        // Verify the builder pattern works
-        let html = r#"<!DOCTYPE html>
-<html>
-<body>
-    <svg xmlns="http://www.w3.org/2000/svg">
-        <rect/>
-    </svg>
-</body>
-</html>"#;
-
-        let document = parse_html().one(html);
-        let selectors = Selectors::compile_with_context("svg|rect", &context).unwrap();
-        let rects = selectors
-            .filter(document.descendants().elements())
-            .collect::<Vec<_>>();
-        assert_eq!(rects.len(), 1);
-    }
-
-    #[test]
-    fn select() {
-        let html = r"
-<title>Test case</title>
-<p class=foo>Foo
-<p>Bar
-<p class=foo>Foo
-";
-
-        let document = parse_html().one(html);
-        let matching = document.select("p.foo").unwrap().collect::<Vec<_>>();
-        assert_eq!(matching.len(), 2);
-        let child = matching[0].as_node().first_child().unwrap();
-        assert_eq!(&**child.as_text().unwrap().borrow(), "Foo\n");
-        assert_eq!(matching[0].attributes.borrow().get("class"), Some("foo"));
-        assert_eq!(
-            matching[0].attributes.borrow().get(local_name!("class")),
-            Some("foo")
-        );
-
-        let selectors = Selectors::compile("p.foo").unwrap();
-        let matching2 = selectors
-            .filter(document.descendants().elements())
-            .collect::<Vec<_>>();
-        assert_eq!(matching, matching2);
-    }
-
-    #[test]
-    fn select_first() {
-        let html = r"
-<title>Test case</title>
-<p class=foo>Foo
-<p>Bar
-<p class=foo>Baz
-";
-
-        let document = parse_html().one(html);
-        let matching = document.select_first("p.foo").unwrap();
-        let child = matching.as_node().first_child().unwrap();
-        assert_eq!(&**child.as_text().unwrap().borrow(), "Foo\n");
-        assert_eq!(matching.attributes.borrow().get("class"), Some("foo"));
-        assert_eq!(
-            matching.attributes.borrow().get(local_name!("class")),
-            Some("foo")
-        );
-
-        assert!(document.select_first("p.bar").is_err());
-    }
-
-    #[test]
-    fn specificity() {
-        let selectors = Selectors::compile(".example, :first-child, div").unwrap();
-        let specificities = selectors
-            .0
-            .iter()
-            .map(|s| s.specificity())
-            .collect::<Vec<_>>();
-        assert_eq!(specificities.len(), 3);
-        assert!(specificities[0] == specificities[1]);
-        assert!(specificities[0] > specificities[2]);
-        assert!(specificities[1] > specificities[2]);
-    }
-}
