+++ src/select/element_impl.rs
@@ -1,0 +1,203 @@
+use super::{AttrValue, BrikSelectors, LocalNameSelector, PseudoClass, PseudoElement};
+use crate::attributes::ExpandedName;
+use crate::iter::NodeIterator;
+use crate::node_data_ref::NodeDataRef;
+use crate::tree::{ElementData, Node, NodeData, NodeRef};
+use html5ever::{local_name, ns, LocalName, Namespace};
+use selectors::attr::{AttrSelectorOperation, CaseSensitivity, NamespaceConstraint};
+use selectors::{matching, OpaqueElement};
+
+/// The definition of whitespace per CSS Selectors Level 3 § 4.
+///
+/// Copied from rust-selectors.
+pub(super) static SELECTOR_WHITESPACE: &[char] = &[' ', '\t', '\n', '\r', '\x0C'];
+
+impl selectors::Element for NodeDataRef<ElementData> {
+    type Impl = BrikSelectors;
+
+    #[inline]
+    fn opaque(&self) -> OpaqueElement {
+        let node: &Node = self.as_node();
+        OpaqueElement::new(node)
+    }
+
+    #[inline]
+    fn is_html_slot_element(&self) -> bool {
+        false
+    }
+    #[inline]
+    fn parent_node_is_shadow_root(&self) -> bool {
+        false
+    }
+    #[inline]
+    fn containing_shadow_host(&self) -> Option<Self> {
+        None
+    }
+
+    #[inline]
+    fn parent_element(&self) -> Option<Self> {
+        self.as_node().parent().and_then(NodeRef::into_element_ref)
+    }
+    #[inline]
+    fn prev_sibling_element(&self) -> Option<Self> {
+        self.as_node().preceding_siblings().elements().next()
+    }
+    #[inline]
+    fn next_sibling_element(&self) -> Option<Self> {
+        self.as_node().following_siblings().elements().next()
+    }
+    #[inline]
+    fn first_element_child(&self) -> Option<Self> {
+        self.as_node().children().elements().next()
+    }
+    #[inline]
+    fn is_empty(&self) -> bool {
+        self.as_node().children().all(|child| match *child.data() {
+            NodeData::Element(_) => false,
+            NodeData::Text(ref text) => text.borrow().is_empty(),
+            _ => true,
+        })
+    }
+    #[inline]
+    fn is_root(&self) -> bool {
+        match self.as_node().parent() {
+            None => false,
+            Some(parent) => matches!(*parent.data(), NodeData::Document(_)),
+        }
+    }
+
+    #[inline]
+    fn is_html_element_in_html_document(&self) -> bool {
+        // FIXME: Have a notion of HTML document v.s. XML document?
+        self.name.ns == ns!(html)
+    }
+
+    #[inline]
+    fn has_local_name(&self, name: &LocalName) -> bool {
+        self.name.local == *name
+    }
+    #[inline]
+    fn has_namespace(&self, namespace: &Namespace) -> bool {
+        self.name.ns == *namespace
+    }
+
+    #[inline]
+    fn is_part(&self, _name: &LocalNameSelector) -> bool {
+        false
+    }
+
+    #[inline]
+    fn imported_part(&self, _: &LocalNameSelector) -> Option<LocalNameSelector> {
+        None
+    }
+
+    #[inline]
+    fn is_pseudo_element(&self) -> bool {
+        false
+    }
+
+    #[inline]
+    fn is_same_type(&self, other: &Self) -> bool {
+        self.name == other.name
+    }
+
+    #[inline]
+    fn is_link(&self) -> bool {
+        self.name.ns == ns!(html)
+            && matches!(
+                self.name.local,
+                local_name!("a") | local_name!("area") | local_name!("link")
+            )
+            && self
+                .attributes
+                .borrow()
+                .map
+                .contains_key(&ExpandedName::new(ns!(), local_name!("href")))
+    }
+
+    #[inline]
+    fn has_id(&self, id: &LocalNameSelector, case_sensitivity: CaseSensitivity) -> bool {
+        self.attributes
+            .borrow()
+            .get(local_name!("id"))
+            .is_some_and(|id_attr| case_sensitivity.eq(id.as_bytes(), id_attr.as_bytes()))
+    }
+
+    #[inline]
+    fn has_class(&self, name: &LocalNameSelector, case_sensitivity: CaseSensitivity) -> bool {
+        let name = name.as_bytes();
+        !name.is_empty()
+            && if let Some(class_attr) = self.attributes.borrow().get(local_name!("class")) {
+                class_attr
+                    .split(SELECTOR_WHITESPACE)
+                    .any(|class| case_sensitivity.eq(class.as_bytes(), name))
+            } else {
+                false
+            }
+    }
+
+    #[inline]
+    fn attr_matches(
+        &self,
+        ns: &NamespaceConstraint<&Namespace>,
+        local_name: &LocalNameSelector,
+        operation: &AttrSelectorOperation<&AttrValue>,
+    ) -> bool {
+        let attrs = self.attributes.borrow();
+        match *ns {
+            NamespaceConstraint::Any => attrs.map.iter().any(|(name, attr)| {
+                name.local == **local_name && operation.eval_str(attr.value.as_str())
+            }),
+            NamespaceConstraint::Specific(ns_url) => attrs
+                .map
+                .get(&ExpandedName::new(ns_url, (**local_name).clone()))
+                .is_some_and(|attr| operation.eval_str(attr.value.as_str())),
+        }
+    }
+
+    fn match_pseudo_element(
+        &self,
+        pseudo: &PseudoElement,
+        _context: &mut matching::MatchingContext<BrikSelectors>,
+    ) -> bool {
+        match *pseudo {}
+    }
+
+    fn match_non_ts_pseudo_class(
+        &self,
+        pseudo: &PseudoClass,
+        _context: &mut matching::MatchingContext<BrikSelectors>,
+    ) -> bool {
+        use self::PseudoClass::*;
+        match *pseudo {
+            Active | Focus | Hover | Enabled | Disabled | Checked | Indeterminate | Visited => {
+                false
+            }
+            AnyLink | Link => {
+                self.name.ns == ns!(html)
+                    && matches!(
+                        self.name.local,
+                        local_name!("a") | local_name!("area") | local_name!("link")
+                    )
+                    && self.attributes.borrow().contains(local_name!("href"))
+            }
+        }
+    }
+
+    #[inline]
+    fn apply_selector_flags(&self, _flags: matching::ElementSelectorFlags) {
+        // No-op for static DOM
+    }
+
+    #[inline]
+    fn has_custom_state(&self, _name: &LocalNameSelector) -> bool {
+        // Brik is a static DOM, no custom states
+        false
+    }
+
+    #[inline]
+    fn add_element_unique_hashes(&self, filter: &mut selectors::bloom::BloomFilter) -> bool {
+        let _ = filter; // Silence unused warning
+        false
+    }
+}
