+++ src/select/specificity.rs
@@ -1,0 +1,8 @@
+/// The specificity of a selector.
+///
+/// Opaque, but ordered.
+///
+/// Determines precedence in the cascading algorithm.
+/// When equal, a rule later in source order takes precedence.
+#[derive(Copy, Clone, Hash, Eq, PartialEq, Ord, PartialOrd)]
+pub struct Specificity(pub(super) u32);
